support -p <python-exe>
Use case, if you need python3 but don't have pyenv and need to refer to
it with just "python3"

Eliminate ad hoc arg parsing and go with getopts. That means we lose
long options. Want to stick with bash builtin, so not using getopt which
might not be there, and might be broken.

Update README, and switch to reference style links.